Move ImageList component to new svelte file
